+++ src/runtime/StateSerialization/ClassManagerState.cs
@@ -6,6 +6,6 @@ namespace Python.Runtime.StateSerialization;
 [Serializable]
 internal class ClassManagerState
 {
-    public Dictionary<PyType, InterDomainContext> Contexts { get; set; }
-    public Dictionary<MaybeType, PyType> Cache { get; set; }
+    public Dictionary<ReflectedClrType, InterDomainContext> Contexts { get; set; }
+    public Dictionary<MaybeType, ReflectedClrType> Cache { get; set; }
 }